完成MCP服务定价功能和用户资金管理模块，修改数据库表结构定义

* boliang: (16 commits)
  完成价格标定
  Merge #11 into boliang from shoubao
  feat: 修改价格字段为整型
  feat:完成MCP价格管理模块开发，支持单次/批量修改以及详细审计日志
  fix: 修复在管理员修改价格时记录应为username
  feat: 完成星量用户资金管理模块开发
  feat:管理员充值时新增充值类型和备注功能
  fix: 修复用户资金管理鉴权与操作人记录
  fix: 调整充值类型下拉框大小与箭头位置
  feat: 资金变动明细增加分页与页数选择
  fix: 修复登录接口表名引用错误并重构查询逻辑
  fix: 修复ServiceBatchUpdatePriceReq结构体 JSON 标签错误
  fix(资金): 金额格式化适配 8 位小数精度
  fix: 服务价格支持小数与批量价格可选参数
  fix: 修复与测试环境数据库字段不对应问题
  fix: 修复资金变动记录写入与排序字段

Merged-by: 陈波良 <accounts_696d8c315e7217876d175a52@mail.teambition.com>

CR-link: https://codeup.aliyun.com/6808bdf3ca2dd04ff7aa3e13/XNetApp/AgentEarth-Mgr/change/13

+++ models/mcp/aeMcpServicesStatisticModel_gen.go
@@ -44,7 +44,7 @@ type (
 		Month        int64     `db:"month"`         // 月
 		Day          int64     `db:"day"`           // 日
 		Hour         int64     `db:"hour"`          // 小时
-		ResponseTime int64     `db:"response_time"` // 小时平均响应时间(毫秒)
+		ResponseTime float64   `db:"response_time"` // 小时平均响应时间(毫秒)
 		RequestTotal int64     `db:"request_total"` // 小时内请求次数
 		CreateTime   time.Time `db:"create_time"`   // 创建时间
 	}
